+++ compiler-tests/src/test/data/diagnostic/inject/InjectingAPlainObjectIsSus.kt
@@ -1,0 +1,5 @@
+// RENDER_DIAGNOSTICS_FULL_TEXT
+object ObjectClass
+
+@Inject
+class Example(obj: <!SUSPICIOUS_OBJECT_INJECTION_WARNING!>ObjectClass<!>)